+++ OpenGLConsoleApp/Light.h
@@ -9,7 +9,7 @@ public:
   Light();
   Light(GLfloat red, GLfloat green, GLfloat blue, GLfloat aIntensity, GLfloat xDirection, GLfloat yDirection, GLfloat zDirection, GLfloat dIntensity);
 
-  void UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColorLocation, GLfloat diffuseIntensityLocation, GLfloat directionLocation);
+  void UseLight(GLuint ambientIntensityLocation, GLuint ambientColorLocation, GLuint diffuseIntensityLocation, GLuint directionLocation);
 
   ~Light();
